+++ go/engine/login_provision.go
@@ -143,10 +143,19 @@ func (e *LoginProvision) device(ctx *Context) error {
 		if err != nil {
 			// could cancel provisionee run here?
 			e.G().Log.Warning("DisplayAndPromptSecret error: %s", err)
-		} else if receivedSecret != nil && len(receivedSecret) > 0 {
+		} else if receivedSecret.Secret != nil && len(receivedSecret.Secret) > 0 {
+			e.G().Log.Debug("received secret, adding to provisionee")
 			var ks kex2.Secret
-			copy(ks[:], receivedSecret)
+			copy(ks[:], receivedSecret.Secret)
 			provisionee.AddSecret(ks)
+		} else if len(receivedSecret.Phrase) > 0 {
+			e.G().Log.Debug("received secret phrase, adding to provisionee")
+			ks, err := libkb.NewKex2SecretFromPhrase(receivedSecret.Phrase)
+			if err != nil {
+				e.G().Log.Warning("DisplayAndPromptSecret error: %s", err)
+			} else {
+				provisionee.AddSecret(ks.Secret())
+			}
 		}
 	}()
 
@@ -209,18 +218,21 @@ func (e *LoginProvision) gpg(ctx *Context) error {
 
 // paper attempts to provision the device via a paper key.
 func (e *LoginProvision) paper(ctx *Context) error {
-	// prompt for the username (if not provided) and load the user:
-	var err error
-	e.user, err = e.loadUser(ctx)
+	// get the paper key from the user
+	kp, err := e.getPaperKey(ctx)
 	if err != nil {
 		return err
 	}
 
-	// find a paper key for this user
-	kp, err := findPaperKeys(ctx, e.G(), e.user)
+	e.G().Log.Debug("paper signing key kid: %s", kp.sigKey.GetKID())
+	e.G().Log.Debug("paper encryption key kid: %s", kp.encKey.GetKID())
+
+	// use the KID to find (and load) the user
+	user, err := e.loadUserByKID(kp.sigKey.GetKID())
 	if err != nil {
 		return err
 	}
+	e.user = user
 
 	// found a paper key that can be used for signing
 	e.G().Log.Debug("found paper key match for %s", e.user.GetName())
@@ -229,6 +241,12 @@ func (e *LoginProvision) paper(ctx *Context) error {
 	// It signs this new device with the paper key.
 	var afterLogin = func(lctx libkb.LoginContext) error {
 		ctx.LoginContext = lctx
+
+		// need lksec to store device keys locally
+		if err := e.fetchLKS(ctx, kp.encKey); err != nil {
+			return err
+		}
+
 		if err := e.makeDeviceKeysWithSigner(ctx, kp.sigKey); err != nil {
 			return err
 		}
@@ -239,8 +257,8 @@ func (e *LoginProvision) paper(ctx *Context) error {
 		return nil
 	}
 
-	// need a session to continue to provision
-	return e.G().LoginState().LoginWithPrompt(e.user.GetName(), ctx.LoginUI, ctx.SecretUI, afterLogin)
+	// need a session to continue to provision, login with paper sigKey
+	return e.G().LoginState().LoginWithKey(ctx.LoginContext, e.user, kp.sigKey, afterLogin)
 }
 
 // passphrase attempts to provision the device via username and
@@ -394,7 +412,7 @@ func (e *LoginProvision) ppStream(ctx *Context) (*libkb.PassphraseStream, error)
 	if ctx.LoginContext != nil {
 		cached := ctx.LoginContext.PassphraseStreamCache()
 		if cached == nil {
-			return nil, errors.New("nil PassphraseStreamCache")
+			return nil, errors.New("LoginProvision: ppStream() -> nil PassphraseStreamCache")
 		}
 		return cached.PassphraseStream(), nil
 	}
@@ -659,6 +677,63 @@ func (e *LoginProvision) checkUserByPGPFingerprint(ctx *Context, fp *libkb.PGPFi
 		return err
 	}
 
+	return nil
+}
+
+func (e *LoginProvision) getPaperKey(ctx *Context) (*keypair, error) {
+	passphrase, err := ctx.SecretUI.GetPaperKeyPassphrase(keybase1.GetPaperKeyPassphraseArg{})
+	if err != nil {
+		return nil, err
+	}
+	paperPhrase := libkb.NewPaperKeyPhrase(passphrase)
+	if paperPhrase.Version() != libkb.PaperKeyVersion {
+		e.G().Log.Debug("paper version mismatch:  generated paper key version = %d, libkb version = %d", paperPhrase.Version(), libkb.PaperKeyVersion)
+		return nil, libkb.KeyVersionError{}
+	}
+
+	bkarg := &PaperKeyGenArg{
+		Passphrase: libkb.NewPaperKeyPhrase(passphrase),
+		SkipPush:   true,
+	}
+	bkeng := NewPaperKeyGen(bkarg, e.G())
+	if err := RunEngine(bkeng, ctx); err != nil {
+		return nil, err
+	}
+
+	return &keypair{sigKey: bkeng.SigKey(), encKey: bkeng.EncKey()}, nil
+}
+
+func (e *LoginProvision) loadUserByKID(kid keybase1.KID) (*libkb.User, error) {
+	arg := libkb.APIArg{
+		Endpoint:     "key/owner",
+		NeedSession:  false,
+		Contextified: libkb.NewContextified(e.G()),
+		Args:         libkb.HTTPArgs{"kid": libkb.S{Val: kid.String()}},
+	}
+	res, err := e.G().API.Get(arg)
+	if err != nil {
+		return nil, err
+	}
+	suid, err := res.Body.AtPath("uid").GetString()
+	if err != nil {
+		return nil, err
+	}
+	uid, err := keybase1.UIDFromString(suid)
+	if err != nil {
+		return nil, err
+	}
+	e.G().Log.Debug("key/owner result uid: %s", uid)
+	loadArg := libkb.NewLoadUserArg(e.G())
+	loadArg.UID = uid
+	return libkb.LoadUser(loadArg)
+}
+
+func (e *LoginProvision) fetchLKS(ctx *Context, encKey libkb.GenericKey) error {
+	gen, clientLKS, err := fetchLKS(ctx, e.G(), encKey)
+	if err != nil {
+		return err
+	}
+	e.lks = libkb.NewLKSecWithClientHalf(clientLKS, gen, e.user.GetUID(), e.G())
 	return nil
 }