+++ go/libkb/account.go
@@ -126,10 +126,12 @@ func (a *Account) CreateStreamCache(tsec *triplesec.Cipher, pps *PassphraseStrea
 
 // SetStreamGeneration sets the passphrase generation on the cached stream
 // if it exists, and otherwise will wind up warning of a problem.
-func (a *Account) SetStreamGeneration(gen PassphraseGeneration) {
+func (a *Account) SetStreamGeneration(gen PassphraseGeneration, nilPPStreamOK bool) {
 	ps := a.PassphraseStreamRef()
 	if ps == nil {
-		a.G().Log.Warning("Passphrase stream was nil; unexpected")
+		if !nilPPStreamOK {
+			a.G().Log.Warning("Passphrase stream was nil; unexpected")
+		}
 	} else {
 		ps.SetGeneration(gen)
 	}